Перенес все (кроме checkAuth и useEffect потом посмотрю что с ними делать) что касется модального окна в ModalWindow.js пробросив пропсы. И оно рабоатет. Дальше нужно будет useState'ы вероятно перенести в родительский компонент, чтобы использовать в RecipesMain.js, чтобы сделать рабочей ссылку "Зарегестрируйтесь".
